app_streamlit/requirements.txt
# CUIDAR IA - Requirements
# Proyecto: Policy Copilot para prevención del suicidio

# === Core Dependencies ===
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1

# === Data Science & Visualization ===
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
openpyxl>=3.1.0 
python-dateutil>=2.8.0

# === LangChain & RAG ===
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-anthropic==0.1.1

# === Vector Store & Embeddings ===
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf==3.17.4
pypdf2==3.0.1

# === LLM Providers ===
openai==1.6.1
anthropic==0.8.1
ollama==0.1.6

# === PDF Generation ===
reportlab==4.0.8
fpdf2==2.7.7
weasyprint==60.2

# === Streamlit (para deploy futuro) ===
streamlit==1.29.0
streamlit-aggrid==0.3.4.post3

# === Utils ===
python-dotenv==1.0.0
pydantic==2.5.3
tiktoken==0.5.2

# === Evaluador de Bases de Datos ===
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0
scikit-learn>=1.3.0
openpyxl>=3.1.0

# === Geocodificación ===
geopy>=2.4.0
urllib3>=2.0.0
requests>=2.31.0

# === Detección de PII ===
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0